pageA/requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
